Move MyHomePage -> own file ('home_page.dart')
