+++ lib/main.dart
@@ -1,7 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:google_fonts/google_fonts.dart';
-import 'dog.dart';
-import 'dog_detail.dart';
+import 'package:flutter_dogs/home_page.dart';
 
 void main() {
   runApp(const DogApp());
@@ -31,70 +29,7 @@ class DogApp extends StatelessWidget {
           secondary: Colors.white,
         ),
       ),
-      home: const MyHomePage(title: 'Dog Gallery'),
-    );
-  }
-}
-
-class MyHomePage extends StatefulWidget {
-  const MyHomePage({Key? key, required this.title}) : super(key: key);
-
-  final String title;
-
-  @override
-  State<MyHomePage> createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(widget.title),
-      ),
-      body: SafeArea(
-        child: ListView.builder(
-          itemCount: Dog.finnishBreeds.length,
-          itemBuilder: (BuildContext context, int index) {
-            return GestureDetector(
-              onTap: () {
-                Navigator.push(
-                  context,
-                  MaterialPageRoute(
-                    builder: (context) {
-                      return DogDetail(dog: Dog.finnishBreeds[index]);
-                    },
-                  ),
-                );
-              },
-              child: buildDogCard(Dog.finnishBreeds[index]),
-            );
-          },
-        ),
-      ),
-    );
-  }
-
-  Widget buildDogCard(Dog dog) {
-    return Card(
-      elevation: 4.0,
-      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(20.0)),
-      child: Padding(
-        padding: const EdgeInsets.all(24.0),
-        child: Column(
-          children: <Widget>[
-            Image(image: AssetImage(dog.imageUrl)),
-            const SizedBox(height: 21.0),
-            Text(
-              dog.breedName,
-              style: GoogleFonts.andada(
-                fontSize: 20.0,
-                fontWeight: FontWeight.bold,
-              ),
-            ),
-          ],
-        ),
-      ),
+      home: const HomePage(title: 'Dog Gallery'),
     );
   }
 }
